+++ crates/core/project-graph/tests/project_graph_test.rs
@@ -1,7 +1,8 @@
 use moon::{generate_project_graph, load_workspace_from};
 use moon_common::Id;
 use moon_config2::{
-    DependencyScope, NodeConfig, RustConfig, ToolchainConfig, WorkspaceConfig, WorkspaceProjects,
+    DependencyScope, PartialConstraintsConfig, PartialNodeConfig, PartialRustConfig,
+    PartialToolchainConfig, PartialWorkspaceConfig, WorkspaceProjects,
 };
 use moon_project::{Project, ProjectDependency, ProjectDependencySource};
 use moon_project_graph::ProjectGraph;
@@ -42,14 +43,14 @@ async fn get_aliases_graph() -> (ProjectGraph, Sandbox) {
 }
 
 async fn get_dependencies_graph(enable_git: bool) -> (ProjectGraph, Sandbox) {
-    let workspace_config = WorkspaceConfig {
-        projects: WorkspaceProjects::Sources(FxHashMap::from_iter([
+    let workspace_config = PartialWorkspaceConfig {
+        projects: Some(WorkspaceProjects::Sources(FxHashMap::from_iter([
             ("a".into(), "a".to_owned()),
             ("b".into(), "b".to_owned()),
             ("c".into(), "c".to_owned()),
             ("d".into(), "d".to_owned()),
-        ])),
-        ..WorkspaceConfig::default()
+        ]))),
+        ..PartialWorkspaceConfig::default()
     };
 
     let sandbox = create_sandbox_with_config(
@@ -70,14 +71,14 @@ async fn get_dependencies_graph(enable_git: bool) -> (ProjectGraph, Sandbox) {
 }
 
 async fn get_dependents_graph() -> (ProjectGraph, Sandbox) {
-    let workspace_config = WorkspaceConfig {
-        projects: WorkspaceProjects::Sources(FxHashMap::from_iter([
+    let workspace_config = PartialWorkspaceConfig {
+        projects: Some(WorkspaceProjects::Sources(FxHashMap::from_iter([
             ("a".into(), "a".to_owned()),
             ("b".into(), "b".to_owned()),
             ("c".into(), "c".to_owned()),
             ("d".into(), "d".to_owned()),
-        ])),
-        ..WorkspaceConfig::default()
+        ]))),
+        ..PartialWorkspaceConfig::default()
     };
 
     let sandbox = create_sandbox_with_config(
@@ -97,21 +98,23 @@ async fn get_tag_constraints_graph<F>(setup: F) -> (ProjectGraph, Sandbox)
 where
     F: FnOnce(&Sandbox),
 {
-    let mut workspace_config = WorkspaceConfig {
-        projects: WorkspaceProjects::Globs(vec!["*".into()]),
-        ..WorkspaceConfig::default()
+    let workspace_config = PartialWorkspaceConfig {
+        projects: Some(WorkspaceProjects::Globs(vec!["*".into()])),
+        constraints: Some(PartialConstraintsConfig {
+            tag_relationships: Some(FxHashMap::from_iter([
+                (
+                    "warrior".into(),
+                    vec![Id::raw("barbarian"), Id::raw("paladin"), Id::raw("druid")],
+                ),
+                (
+                    "mage".into(),
+                    vec![Id::raw("wizard"), Id::raw("sorcerer"), Id::raw("druid")],
+                ),
+            ])),
+            ..PartialConstraintsConfig::default()
+        }),
+        ..PartialWorkspaceConfig::default()
     };
-
-    workspace_config.constraints.tag_relationships = FxHashMap::from_iter([
-        (
-            "warrior".into(),
-            vec![Id::raw("barbarian"), Id::raw("paladin"), Id::raw("druid")],
-        ),
-        (
-            "mage".into(),
-            vec![Id::raw("wizard"), Id::raw("sorcerer"), Id::raw("druid")],
-        ),
-    ]);
 
     let sandbox = create_sandbox_with_config(
         "project-graph/tag-constraints",
@@ -132,14 +135,14 @@ async fn get_type_constraints_graph<F>(setup: F) -> (ProjectGraph, Sandbox)
 where
     F: FnOnce(&Sandbox),
 {
-    let mut workspace_config = WorkspaceConfig {
-        projects: WorkspaceProjects::Globs(vec!["*".into()]),
-        ..WorkspaceConfig::default()
+    let workspace_config = PartialWorkspaceConfig {
+        projects: Some(WorkspaceProjects::Globs(vec!["*".into()])),
+        constraints: Some(PartialConstraintsConfig {
+            enforce_project_type_relationships: Some(true),
+            ..PartialConstraintsConfig::default()
+        }),
+        ..PartialWorkspaceConfig::default()
     };
-
-    workspace_config
-        .constraints
-        .enforce_project_type_relationships = true;
 
     let sandbox = create_sandbox_with_config(
         "project-graph/type-constraints",
@@ -157,15 +160,15 @@ where
 }
 
 async fn get_queries_graph() -> (ProjectGraph, Sandbox) {
-    let workspace_config = WorkspaceConfig {
-        projects: WorkspaceProjects::Globs(vec!["*".into()]),
-        ..WorkspaceConfig::default()
+    let workspace_config = PartialWorkspaceConfig {
+        projects: Some(WorkspaceProjects::Globs(vec!["*".into()])),
+        ..PartialWorkspaceConfig::default()
     };
 
-    let toolchain_config = ToolchainConfig {
-        node: Some(NodeConfig::default()),
-        rust: Some(RustConfig::default()),
-        ..ToolchainConfig::default()
+    let toolchain_config = PartialToolchainConfig {
+        node: Some(PartialNodeConfig::default()),
+        rust: Some(PartialRustConfig::default()),
+        ..PartialToolchainConfig::default()
     };
 
     let sandbox = create_sandbox_with_config(
@@ -183,15 +186,15 @@ async fn get_queries_graph() -> (ProjectGraph, Sandbox) {
 
 #[tokio::test]
 async fn can_use_map_and_globs_setting() {
-    let workspace_config = WorkspaceConfig {
-        projects: WorkspaceProjects::Both {
+    let workspace_config = PartialWorkspaceConfig {
+        projects: Some(WorkspaceProjects::Both {
             globs: string_vec!["deps/*"],
             sources: FxHashMap::from_iter([
                 ("basic".into(), "basic".to_owned()),
                 ("noConfig".into(), "no-config".to_owned()),
             ]),
-        },
-        ..WorkspaceConfig::default()
+        }),
+        ..PartialWorkspaceConfig::default()
     };
 
     let sandbox = create_sandbox_with_config("projects", Some(workspace_config), None, None);
@@ -223,12 +226,12 @@ async fn can_use_map_and_globs_setting() {
 
 #[tokio::test]
 async fn can_generate_with_deps_cycles() {
-    let workspace_config = WorkspaceConfig {
-        projects: WorkspaceProjects::Sources(FxHashMap::from_iter([
+    let workspace_config = PartialWorkspaceConfig {
+        projects: Some(WorkspaceProjects::Sources(FxHashMap::from_iter([
             ("a".into(), "a".to_owned()),
             ("b".into(), "b".to_owned()),
-        ])),
-        ..WorkspaceConfig::default()
+        ]))),
+        ..PartialWorkspaceConfig::default()
     };
 
     let sandbox =
@@ -311,9 +314,9 @@ mod globs {
 
     #[tokio::test]
     async fn ignores_moon_dot_folder() {
-        let workspace_config = WorkspaceConfig {
-            projects: WorkspaceProjects::Globs(string_vec!["*"]),
-            ..WorkspaceConfig::default()
+        let workspace_config = PartialWorkspaceConfig {
+            projects: Some(WorkspaceProjects::Globs(string_vec!["*"])),
+            ..PartialWorkspaceConfig::default()
         };
 
         // Use git so we can test against the .git folder
@@ -335,9 +338,9 @@ mod globs {
 
     #[tokio::test]
     async fn filters_ignored_sources() {
-        let workspace_config = WorkspaceConfig {
-            projects: WorkspaceProjects::Globs(string_vec!["*"]),
-            ..WorkspaceConfig::default()
+        let workspace_config = PartialWorkspaceConfig {
+            projects: Some(WorkspaceProjects::Globs(string_vec!["*"])),
+            ..PartialWorkspaceConfig::default()
         };
 
         // Use git so we can test against the .git folder
@@ -360,9 +363,9 @@ mod globs {
 
     #[tokio::test]
     async fn supports_all_id_formats() {
-        let workspace_config = WorkspaceConfig {
-            projects: WorkspaceProjects::Globs(string_vec!["*"]),
-            ..WorkspaceConfig::default()
+        let workspace_config = PartialWorkspaceConfig {
+            projects: Some(WorkspaceProjects::Globs(string_vec!["*"])),
+            ..PartialWorkspaceConfig::default()
         };
 
         let sandbox =
@@ -442,14 +445,14 @@ mod to_dot {
 
     #[tokio::test]
     async fn renders_partial_tree() {
-        let workspace_config = WorkspaceConfig {
-            projects: WorkspaceProjects::Sources(FxHashMap::from_iter([
+        let workspace_config = PartialWorkspaceConfig {
+            projects: Some(WorkspaceProjects::Sources(FxHashMap::from_iter([
                 ("a".into(), "a".to_owned()),
                 ("b".into(), "b".to_owned()),
                 ("c".into(), "c".to_owned()),
                 ("d".into(), "d".to_owned()),
-            ])),
-            ..WorkspaceConfig::default()
+            ]))),
+            ..PartialWorkspaceConfig::default()
         };
 
         let sandbox = create_sandbox_with_config(
