+++ crates/deno/platform/Cargo.toml
@@ -26,8 +26,3 @@ rustc-hash = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 tokio = { workspace = true }
-
-[dev-dependencies]
-moon = { path = "../../core/moon" }
-moon_project_graph = { path = "../../core/project-graph" }
-moon_test_utils = { path = "../../core/test-utils" }
