+++ nextgen/config/src/workspace/constraints_config.rs
@@ -1,9 +1,8 @@
 use moon_common::Id;
 use rustc_hash::FxHashMap;
 use schematic::Config;
-use serde::Serialize;
 
-#[derive(Config, Serialize)]
+#[derive(Config)]
 pub struct ConstraintsConfig {
     #[setting(default = true)]
     pub enforce_project_type_relationships: bool,
